requirements.txt
asttokens==2.4.1
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
greenlet==3.0.3
ipykernel==6.29.0
ipython==8.20.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib-inline==0.1.6
mysql-connector-python==8.3.0
nest-asyncio==1.5.9
numpy==1.26.3
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.2
six==1.16.0
SQLAlchemy==2.0.25
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
wcwidth==0.2.13
